+++ j2e/packageStorage/src/main/java/fr/polytech/team/n/PackageInformation.java
@@ -9,5 +9,5 @@ public interface PackageInformation {
 
     List<Package> getPackages();
 
-    Package getPackageById(int id);
+    Package getPackageById(Long id);
 }
